requirements.txt
streamlit==1.38.0
torch>=2.2
torchvision>=0.17
open_clip_torch>=2.24
scikit-learn>=1.4
Pillow>=10.3
numpy>=1.26